+++ src/types.ts
@@ -15,8 +15,8 @@ export type Action = (payload: any) => ActionGeneratorResult;
 
 export interface StoreDescriptor {
     state: any;
-    mutations: Record<string, MutationFunction<any>>;
-    getters: Record<string, any>;
+    mutations?: Record<string, MutationFunction<any>>;
+    getters?: Record<string, any>;
     modules?: Record<string, StoreDescriptor>;
 }